+++ Kernel/Bus/VirtIO/VirtIO.cpp
@@ -5,11 +5,11 @@
  */
 
 #include <Kernel/Bus/PCI/IDs.h>
+#include <Kernel/Bus/VirtIO/VirtIO.h>
+#include <Kernel/Bus/VirtIO/VirtIOConsole.h>
+#include <Kernel/Bus/VirtIO/VirtIORNG.h>
 #include <Kernel/CommandLine.h>
 #include <Kernel/Sections.h>
-#include <Kernel/VirtIO/VirtIO.h>
-#include <Kernel/VirtIO/VirtIOConsole.h>
-#include <Kernel/VirtIO/VirtIORNG.h>
 
 namespace Kernel {
