+++ Kernel/Bus/VirtIO/VirtIO.h
@@ -9,10 +9,10 @@
 #include <AK/NonnullOwnPtrVector.h>
 #include <Kernel/Bus/PCI/Access.h>
 #include <Kernel/Bus/PCI/Device.h>
+#include <Kernel/Bus/VirtIO/VirtIOQueue.h>
 #include <Kernel/IO.h>
 #include <Kernel/Interrupts/IRQHandler.h>
 #include <Kernel/Memory/MemoryManager.h>
-#include <Kernel/VirtIO/VirtIOQueue.h>
 
 namespace Kernel {
