+++ Kernel/Bus/VirtIO/VirtIORNG.cpp
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
+#include <Kernel/Bus/VirtIO/VirtIORNG.h>
 #include <Kernel/Sections.h>
-#include <Kernel/VirtIO/VirtIORNG.h>
 
 namespace Kernel {
